Widen yield timer payload u16 → u32 through handler ABI and RIR
Prep for native ms-precision timers on both backends. The 1-second
TimerWheel path that Step B wired in only ever reads seconds, so until
this commit the 16-bit ms payload (≤65s) was a non-issue. The followup
IORING_OP_TIMEOUT (B2) and epoll min-heap (B3) need precise ms plus a
range wider than 65s — widen the whole pipeline in one go so neither
backend has a different truncation behavior.

- handler_abi.h: make_yield_payload takes u32 and packs into
  (upstream_id << 16 | status_code). New HandlerResult::yield_payload_u32()
  decoder. The status_code + upstream_id slots are unused for Yield
  actions, so re-tasking them is additive-only.
- rir.h Function: yield_payload u16* → u32*. rir_builder set_yield_payload
  takes const u32*.
- codegen.cc make_result_yield: takes u32 payload; state-machine
  emission reads u32 from fn.yield_payload.
- lower_rir.cc: propagates mir wait.ms as u32.
- analyze.cc: drop the 0xFFFF cap, keep the non-negative check.
- jit_dispatch.h: JitDispatchOutcome exposes raw timer_ms (u32);
  timer_seconds_from_ms kept as the bucket helper for the legacy
  TimerWheel path. callbacks_impl handle_jit_outcome bucketizes at the
  call site rather than the dispatch layer, so future callers can
  consume timer_ms unchanged.
- tests: rename the now-invalid "rejects >u16" frontend test to the
  positive case (accepts 100000 ms); new jit_dispatch test verifies
  100000 ms round-trips through the codegen + unpack pipeline.

Build + 20/20 ctest suites pass.

+++ include/rut/compiler/rir.h
@@ -335,10 +335,10 @@ struct Function {
     u32 yield_count;
 
     // Per-yield payload. For a Timer yield, yield_payload[i] is the
-    // duration in milliseconds. Arena-allocated, length = yield_count.
-    // (Other yield kinds will extend this or use a richer structure;
-    // for v1 timer-only it stays a flat u16 array.)
-    u16* yield_payload;
+    // duration in milliseconds (u32 ≈ 49 days). Arena-allocated, length
+    // = yield_count. Other yield kinds will extend this or use a richer
+    // structure; for v1 timer-only it stays a flat u32 array.
+    u32* yield_payload;
 
     Block* entry() { return block_count > 0 ? &blocks[0] : nullptr; }
     const Block* entry() const { return block_count > 0 ? &blocks[0] : nullptr; }
